+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Versions
 
 ## 2.19.0 (??)
-- Feature: new replacement type for unordered dates
+- Feature: new replacement for unordered dates
+- Feature: new replacement for dates with the wrong article
 - Fix: check there actually are fixed replacements before saving the review
 - Fix: remove autofocus on editing custom snippet to improve the size of the frame on mobile devices